Make Stereotype editor not crash
It does not show the stereotype attributes yet, though.
