+++ gaphor/adapters/__init__.py
@@ -9,7 +9,7 @@ import gaphor.adapters.classes.interfaceconnect
 import gaphor.adapters.components.connectorconnect
 import gaphor.adapters.interactions.messageconnect
 import gaphor.adapters.profiles.extensionconnect
-import gaphor.adapters.profiles.stereotypespage
+import gaphor.adapters.profiles.stereotypepage
 import gaphor.adapters.profiles.metaclasseditor
 import gaphor.adapters.usecases.usecaseconnect
