+++ gaphor/adapters/propertypages.py
@@ -1138,7 +1138,7 @@ class LineStylePage(object):
     def __init__(self, item):
         super(LineStylePage, self).__init__()
         self.item = item
-        self.size_group = Gtk.SizeGroup(Gtk.SizeGroupMode.HORIZONTAL)
+        self.size_group = Gtk.SizeGroup(mode=Gtk.SizeGroupMode.HORIZONTAL)
 
     def construct(self):
         page = Gtk.VBox()
@@ -1149,7 +1149,7 @@ class LineStylePage(object):
         self.size_group.add_widget(label)
         hbox.pack_start(label, False, True, 0)
 
-        button = Gtk.CheckButton(_("Orthogonal"))
+        button = Gtk.CheckButton(label=_("Orthogonal"))
         button.set_active(self.item.orthogonal)
         button.connect("toggled", self._on_orthogonal_change)
         hbox.pack_start(button, True, True, 0)
@@ -1166,7 +1166,7 @@ class LineStylePage(object):
         self.size_group.add_widget(label)
         hbox.pack_start(label, False, True, 0)
 
-        button = Gtk.CheckButton(_("Horizontal"))
+        button = Gtk.CheckButton(label=_("Horizontal"))
         button.set_active(self.item.horizontal)
         button.connect("toggled", self._on_horizontal_change)
         hbox.pack_start(button, True, True, 0)
